+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fix non-resolved relative path when importing EXMARaLDA files. 
+- Limit the table width when listing the module properties, so they fit in the
+  current terminal.
 
 ## [0.7.0] - 2024-05-23